Added interface to push data to Cassandra
Added interface to push data to Cassandra. The behavior is worked out and the implementation is in progress.
This code can be added to PySpark code to pull data from Arrow as well.
The idea of this code is to turn it into a daemon that can read from Arrow, generate a Cassandra schema, and bulk load time-series data into Cassandra, and metadata into Mongo.