Added precision_test
Precision test reads the PHM08 values from Plasma and generates a DataFrame from them.
It loads a copy of the original data (slightly modified to use comma-delimiters and include a header) into a DataFrame as well.
It then compares the two DataFrames in a flexible way to ensure that data wasn't lost in the workflow, the number of columns was retained, and any changes in value due to using precision floating point data types is determined.